backend/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
python-multipart
segmentation-models-pytorch
scikit-learn
scikit-image
plotly
python-dotenv
